Wyse QC Excel Macro v 3.2
Billing Quality Control checks and Excel Spreadsheet formatting macro.  Read data from the property-specific billing setup spreadsheet from prep and compared to the Billing Summary Spreadsheet generated at the end of the automated billing process.  Worked for submetered properties with one or more utilities, with one or more meters per utility, with multiple tiers of rate per utility.

This code was specific to the Wyse billing department at Conservice before that department was dissolved.  It is an abandoned project.